+++ src/EventSourcing.Commands/CommandExtensions.cs
@@ -7,34 +7,34 @@ public static class CommandExtensions
 {
     public static CommandResult<TFailure>.Processed_ ToFailureResult<TFailure>(this Command command, TFailure failure, string resultMessage)
         where TFailure : IFailure<TFailure>
-        => command.ToEmptyProcessingResult<TFailure>(resultMessage, FunctionalResult<TFailure>.Failed(failure));
+        => command.ToEmptyProcessingResult(resultMessage, FunctionalResult<TFailure>.Failed(failure));
     
     public static CommandResult<TFailure>.Processed_ ToEmptyProcessingResult<TFailure>(this Command command, string resultMessage, FunctionalResult<TFailure>? functionalResult = null)
         where TFailure : IFailure<TFailure>
-        => command.ToProcessedResult<TFailure>(Array.Empty<IEventPayload>(), functionalResult ?? FunctionalResult<TFailure>.Ok(resultMessage));
+        => command.ToProcessedResult(Array.Empty<IEventPayload>(), functionalResult ?? FunctionalResult<TFailure>.Ok(resultMessage));
 
     public static CommandResult<TFailure>.Processed_ ToOkResult<TFailure>(this Command command, IEventPayload resultEvent, string resultMessage)
         where TFailure : IFailure<TFailure>
-        => command.ToProcessedResult<TFailure>([resultEvent], FunctionalResult<TFailure>.Ok(resultMessage));
+        => command.ToProcessedResult([resultEvent], FunctionalResult<TFailure>.Ok(resultMessage));
 
     public static CommandResult<TFailure>.Processed_ ToProcessedResult<TFailure>(this Command command, IEventPayload resultEvent, FunctionalResult<TFailure> functionalResult)
         where TFailure : IFailure<TFailure>
-        => command.ToProcessedResult<TFailure>([resultEvent], functionalResult);
+        => command.ToProcessedResult([resultEvent], functionalResult);
 
     public static CommandResult<TFailure>.Processed_ ToProcessedResult<TFailure>(this Command command,
         IReadOnlyCollection<IEventPayload> resultEvents, FunctionalResult<TFailure> functionalResult)
-        where TFailure : IFailure<TFailure>
-        => new CommandResult<TFailure>.Processed_(resultEvents, command.Id, functionalResult);
+        where TFailure : IFailure<TFailure> =>
+        new(resultEvents, command.Id, functionalResult);
 
     internal static string DefaultOkMessage(this Command command) => $"Successfully processed command {command}";
 }
 
-public static partial class OperationResultExtensions
+public static class ResultExtensions
 {
     public static CommandResult<TFailure> ToProcessedResult<T, TFailure>(this IResult<T, TFailure> operationResult, Command command, string? successMessage = null)
         where T : IEventPayload
         where TFailure : IFailure<TFailure>
-        => operationResult.ToProcessedResult<T, TFailure>(command, successMessage != null ? _ => successMessage : null);
+        => operationResult.ToProcessedResult(command, successMessage != null ? _ => successMessage : null);
 
     public static CommandResult<TFailure>
         ToProcessedResult<T, TFailure>(this IResult<T, TFailure> operationResult, Command command, Func<T, string?>? successMessage)
@@ -42,8 +42,8 @@ public static partial class OperationResultExtensions
         where TFailure : IFailure<TFailure>
         => operationResult
             .Match(
-                ok => command.ToProcessedResult<TFailure>(ok, FunctionalResult<TFailure>.Ok(successMessage?.Invoke(ok) ?? command.DefaultOkMessage())),
-                failure => command.ToFailureResult<TFailure>(failure, failure.Message)
+                ok => command.ToProcessedResult(ok, FunctionalResult<TFailure>.Ok(successMessage?.Invoke(ok) ?? command.DefaultOkMessage())),
+                failure => command.ToFailureResult(failure, failure.Message)
             );
 
     public static CommandResult<TFailure>
@@ -52,8 +52,8 @@ public static partial class OperationResultExtensions
         where TFailure : IFailure<TFailure>
         => operationResult
             .Match(
-                ok => command.ToProcessedResult<TFailure>(ok, FunctionalResult<TFailure>.Ok(successMessage?.Invoke(ok) ?? command.DefaultOkMessage())),
-                failure => command.ToFailureResult<TFailure>(failure, failure.Message)
+                ok => command.ToProcessedResult(ok, FunctionalResult<TFailure>.Ok(successMessage?.Invoke(ok) ?? command.DefaultOkMessage())),
+                failure => command.ToFailureResult(failure, failure.Message)
             );
 
     public static CommandResult<TFailure>
@@ -62,8 +62,8 @@ public static partial class OperationResultExtensions
         where TFailure : IFailure<TFailure>
         => operationResult
             .Match(
-                ok => command.ToProcessedResult<TFailure>(ok.eventPayloads, FunctionalResult<TFailure>.Ok(ok.successMessage)),
-                failure => command.ToFailureResult<TFailure>(failure, failure.Message)
+                ok => command.ToProcessedResult(ok.eventPayloads, FunctionalResult<TFailure>.Ok(ok.successMessage)),
+                failure => command.ToFailureResult(failure, failure.Message)
             );
 
     public static CommandResult<TFailure>
@@ -72,7 +72,7 @@ public static partial class OperationResultExtensions
         where TFailure : IFailure<TFailure>
         => operationResult
             .Match(
-                ok => command.ToProcessedResult<TFailure>(ok.eventPayload, FunctionalResult<TFailure>.Ok(ok.successMessage)),
-                failure => command.ToFailureResult<TFailure>(failure, failure.Message)
+                ok => command.ToProcessedResult(ok.eventPayload, FunctionalResult<TFailure>.Ok(ok.successMessage)),
+                failure => command.ToFailureResult(failure, failure.Message)
             );
 }
