+++ src/EventSourcing.Persistence.EntityFramework.Sqlite/EventSourcing.Persistence.EntityFramework.Sqlite.csproj
@@ -30,7 +30,7 @@
 		<Titel>Bluehands.EventSourcing.Persistence.Sqlite</Titel>
 		<Authors>bluehands</Authors>
 		<PackageDescription>Sqlite persistence for EventSourcing</PackageDescription>
-		<Copyright>Copyright 2024 bluehands GmbH &amp; C.ommunication KG</Copyright>
+		<Copyright>Copyright 2025 bluehands GmbH &amp; C.ommunication KG</Copyright>
 		<PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
 		<PackageLicenseExpression>MIT</PackageLicenseExpression>
 		<!--<PackageIcon>FunicularSwitch.png</PackageIcon>-->
